# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freight Management Platform API
  description: |
    Comprehensive API for Ethiopian freight logistics platform.

    ## Authentication
    Most endpoints require authentication via session cookie or Bearer token.
    - **Web clients**: Use session cookie with CSRF token in X-CSRF-Token header
    - **Mobile clients**: Use Bearer token in Authorization header

    ## Rate Limiting
    All endpoints are rate-limited. When exceeded, returns 429 with Retry-After header.

    ## Error Handling
    - 400: Validation error (Zod schema)
    - 401: Unauthorized
    - 403: Forbidden (insufficient permissions)
    - 404: Resource not found
    - 409: Conflict (state machine violation)
    - 429: Rate limit exceeded
    - 500: Internal server error
  version: 1.0.0
  contact:
    name: Freight Platform Support
    email: support@freightplatform.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000/api
    description: Development
  - url: https://api.freightplatform.com/api
    description: Production
...
paths:
...
  /user/profile:
    get:
      tags: [user]
      summary: Get user profile
      operationId: getUserProfile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      tags: [user]
      summary: Update user profile
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
  schemas:
...
    UserRole:
      type: string
      enum: [SHIPPER, CARRIER, DISPATCHER, ADMIN, SUPER_ADMIN]
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        status:
          type: string
        organizationId:
          type: string
        isEmailVerified:
          type: boolean
        isPhoneVerified:
          type: boolean
    UserProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            organization:
              $ref: "#/components/schemas/OrganizationSummary"
            createdAt:
              type: string
              format: date-time
            lastLoginAt:
              type: string
              format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
...
    OrganizationSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        isVerified:
          type: boolean

    # --------------------------------------------------------------------------
    # LOAD
    # --------------------------------------------------------------------------
...
security:
  - bearerAuth: []
  - cookieAuth: []